+++ src/extension.ts
@@ -12,7 +12,8 @@ export interface PickItem extends QuickPickItem {
     label: string;
     description: string;
     detail: string;
-    fullIssue: any;
+    fullIssue?: any;
+    [key: string]: any;
 }
 
 export function activate(context: vscode.ExtensionContext) {
@@ -23,18 +24,28 @@ export function activate(context: vscode.ExtensionContext) {
         "serverPort": null,
         "serverIsSsl": null,
         "apiKey": null,
-        "rejectUnauthorized": true
+        "rejectUnauthorized": true,
+        "projectName": null
     };
 
     let hadErrors = false;
 
     // TODO: Change this poor settings management system
     let wsSettings: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration('redmine');
-    
+
     let configUpdate = () => {
         hadErrors = false;
         wsSettings = vscode.workspace.getConfiguration('redmine');
         for (let key in settings) {
+            if (key === "projectName") {
+                if (wsSettings.has(key) && wsSettings.get(key) !== "") {
+                    settings[key] = wsSettings.get(key);
+                } else {
+                    settings[key] = null;
+                }
+
+                continue;
+            }
             if (!wsSettings.has(key) || (wsSettings.get(key) === "")) {
                 vscode.window.showErrorMessage(`Redmine integration: ${key} is required`);
                 hadErrors = true;
@@ -72,7 +83,7 @@ export function activate(context: vscode.ExtensionContext) {
                 }
             })).then((issue) => {
                 if (issue === undefined) return;
-                
+
                 let controller = new IssueController(issue.fullIssue, redmine);
 
                 controller.listActions();
@@ -105,13 +116,56 @@ export function activate(context: vscode.ExtensionContext) {
             }
 
             let promise = redmine.getIssueById(issueId);
-    
+
             promise.then((issue) => {
                 if (!issue) return;
-                
+
                 let controller = new IssueController(issue.issue, redmine);
-    
+
                 controller.listActions();
+            }, (error) => {
+                vscode.window.showErrorMessage(error);
+            })
+
+            vscode.window.withProgress({
+                location: vscode.ProgressLocation.Window
+            }, (progress) => {
+                progress.report({ "message": `Waiting for response from ${redmine.host}...` });
+                return promise;
+            });
+        });
+    });
+
+    let newIssue = vscode.commands.registerCommand('redmine.newIssue', () => {
+        if (redmine == null) {
+            vscode.window.showErrorMessage(`Redmine integration: Configuration file is not complete!`);
+            return;
+        }
+
+        const open = (projectName: string) => {
+            vscode.commands.executeCommand("vscode.open", vscode.Uri.parse(`${redmine.url}/projects/${projectName}/issues/new`)).then((success) => {
+            }, (reason) => {
+                vscode.window.showErrorMessage(reason);
+            });
+        }
+
+        if (settings.projectName === null) {
+            let promise = redmine.getProjects();
+    
+            promise.then((projects) => {
+                vscode.window.showQuickPick<PickItem>(projects.projects.map((project) => {
+                    return {
+                        "label": `${project.name}`,
+                        "description": project.description.split("\n").join(" ").split("\r").join(""),
+                        "detail": `${project.identifier}`,
+                        "identifier": project.identifier
+                    }
+                }), {
+                    placeHolder: "Choose project to create issue in"
+                }).then((project) => {
+                    if (project === undefined) return;
+                    open(project.identifier);
+                })
             }, (error) => {
                 vscode.window.showErrorMessage(error);
             })
@@ -122,11 +176,16 @@ export function activate(context: vscode.ExtensionContext) {
                 progress.report({ "message": `Waiting for response from ${redmine.host}...` });
                 return promise;
             });
-        });
+        } else {
+            open(settings.projectName);
+        }
+
+
     });
 
     context.subscriptions.push(listIssues);
     context.subscriptions.push(getIssue);
+    context.subscriptions.push(newIssue);
 }
 
 // this method is called when your extension is deactivated